Rename gateway -> api
