+++ apps/drain-safe/src/components/App.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Title, TextField, Text } from '@gnosis.pm/safe-react-components';
 import { useSafeAppsSDK } from '@gnosis.pm/safe-apps-react-sdk';
 import web3Utils from 'web3-utils';
-import { Asset } from '../utils/gateway';
+import { Asset } from '../utils/api';
 import useBalances from '../hooks/use-balances';
 import { tokenToTx } from '../utils/sdk-helpers';
 import FormContainer from './FormContainer';
